add plugins to extend functionality
This adds plugins which can extend the functionality of the webssh
server. These are particularly useful when the server is intended to be
used headless (i.e. for the backend ssh functionality, with a custom
UI).

To facilitate this usecase, a plugin module is defined, with one plugin
to start: SocketBuilder. SocketBuilder allows an extension to change how
the ssh session is transported. One usecase is to allow it to be
tunneled over something other than a local tcp connection, for example.

Further, since we will want to give clients of the package the ability
to register plugins/etc, expose the core "start" functionality in a
separate method, defauting the normal behaviour of the package to
starting with no plugins.