add option to allow cross origin requests
A usecase that recently arose involves webssh acting as an API for
multiple UIs served in different locations. When run this way, the
current XSRF protections don't really work that well, for a few reasons:

1. The UI being hosted elsewhere means the server-side embedded XSRF
rendering can no longer work.
2. The host itself is somewhat dynamic, meaning that we cannot hardcode
the list of allowed origins easily.

This introduces a flag, off by default, which disables the xsrf
detection entirely.